Move transitivity to its own file
